bots/user_history_extractor/requirements.txt
telethon
PySide6
python-dotenv